Use the "linear" output of tacotron instead of the mel output for inference
This has the advantage of using the decoder postnet (it previously used the
mel output which is the input to this postnet). The fatchord implementation
does not evaluate the postnet loss against the linear spectrogram as
described in 1703.10135, instead it uses the mel-scale spectrogram which
conditions the postnet output to also be mel scale.

The tacotron1 authors state that their use of Griffin-Lim as the vocoder
motivates the use of a linear scale for the postnet output, however a
mel scale is more practical here since WaveRNN is the default vocoder.